Might main print __file__ and sys.argv on exceptions before exiting on them?
Docs say: `sys.exit(code)` `raises SysExit(code)`. `SysExit` is not a subclass
of `Exception`. That means it would not be catched and exits the program using
the passed exit code, so that case should work even without doing anything.

But it looks like the logging in case of Exceptions reaching runpy() could
help to understand the source of such Exceptions better?

Signed-off-by: Bernhard Kaindl <bernhardkaindl7@gmail.com>